requirements.txt

fastapi==0.116.1
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.8.2


httpx==0.27.2
python-dotenv==1.0.1
python-multipart==0.0.5



pydantic-settings==2.2.1